requirements.txt
numpy>=1.22.3
pandas>=1.4.2
scipy>=1.8.0
scikit-learn>=1.0.2
matplotlib>=3.5.1
tensorflow==2.6.0
streamlit==1.8.1
pillow==9.1.0
keras==2.6.0
keras-preprocessing==1.1.2
opencv-python==4.5.5.64